+++ commons/api/src/main/kotlin/com/umpa/commons/api/response/ResultType.kt
@@ -1,4 +1,4 @@
-package com.umpa.response
+package com.umpa.commons.api.response
 
 enum class ResultType {
     SUCCESS, ERROR
