+++ commons/enums/src/main/kotlin/com/umpa/commons/enums/ContentType.kt
@@ -1,4 +1,4 @@
-package com.umpa
+package com.umpa.commons.enums
 
 enum class ContentType {
     PLAYLIST, DAILY, RELAY_PLAYLIST
