+++ storage/db-core/src/main/kotlin/com/umpa/storage/db/core/config/CoreJpaConfig.kt
@@ -1,4 +1,4 @@
-package com.umpa.config
+package com.umpa.storage.db.core.config
 
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
