+++ storage/db-core/src/main/kotlin/com/umpa/storage/db/core/feed/FeedRepository.kt
@@ -1,4 +1,4 @@
-package com.umpa.feed
+package com.umpa.storage.db.core.feed
 
 import org.springframework.data.jpa.repository.JpaRepository
