+++ storage/db-core/src/main/kotlin/com/umpa/storage/db/core/hashtag/HashtagRepository.kt
@@ -1,4 +1,4 @@
-package com.umpa.hashtag
+package com.umpa.storage.db.core.hashtag
 
 import org.springframework.data.jpa.repository.JpaRepository
