+++ storage/db-core/src/main/kotlin/com/umpa/storage/db/core/playlist/PlaylistRepository.kt
@@ -1,4 +1,4 @@
-package com.umpa.playlist
+package com.umpa.storage.db.core.playlist
 
 import org.springframework.data.jpa.repository.JpaRepository
