+++ umpa-core-api/src/main/kotlin/com/umpa/core/domain/content/playlist/PlaylistCreator.kt
@@ -1,7 +1,7 @@
 package com.umpa.core.domain.content.playlist
 
-import com.umpa.playlist.PlaylistEntity
-import com.umpa.playlist.PlaylistRepository
+import com.umpa.storage.db.core.playlist.PlaylistEntity
+import com.umpa.storage.db.core.playlist.PlaylistRepository
 import org.springframework.stereotype.Component
 
 @Component
