+++ umpa-core-api/src/main/kotlin/com/umpa/core/domain/user/UserCreator.kt
@@ -1,7 +1,7 @@
 package com.umpa.core.domain.user
 
-import com.umpa.user.UserEntity
-import com.umpa.user.UserRepository
+import com.umpa.storage.db.core.user.UserEntity
+import com.umpa.storage.db.core.user.UserRepository
 import org.springframework.stereotype.Component
 import java.time.LocalDateTime
