+++ umpa-core-api/src/main/kotlin/com/umpa/core/domain/user/UserValidator.kt
@@ -3,7 +3,7 @@ package com.umpa.core.domain.user
 import com.umpa.core.support.exceptions.CoreApiException
 import com.umpa.core.support.exceptions.ErrorType
 import com.umpa.core.support.utils.BcryptPasswordUtils
-import com.umpa.user.UserRepository
+import com.umpa.storage.db.core.user.UserRepository
 import org.springframework.stereotype.Component
 
 @Component
